requirements.txt
anyio==3.6.1
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==7.1.2
cloudinary==1.29.0
coverage==5.5
coveralls==2.1.2
cryptography==3.4.8
docopt==0.6.2
docutils==0.19
ecdsa==0.18.0
fastapi==0.81.0
# Editable install with no version control (fastapi-jwt-auth==0.5.0)
-e /home/deep/Python Training/Flask Training/fastapi-jwt-auth
flit==3.7.1
flit_core==3.7.1
ghp-import==2.1.0
greenlet==1.1.3
h11==0.9.0
httptools==0.1.2
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
Jinja2==3.1.2
Markdown==3.3.7
markdown-include==0.5.1
MarkupSafe==2.1.1
mergedeep==1.3.4
mkdocs==1.3.1
mkdocs-material==5.5.14
mkdocs-material-extensions==1.0.3
more-itertools==8.14.0
packaging==21.3
passlib==1.7.4
pluggy==0.13.1
psycopg2==2.9.3
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.2
Pygments==2.13.0
PyJWT==1.7.1
pymdown-extensions==9.5
pyparsing==3.0.9
pytest==6.0.1
pytest-cov==2.10.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
pyyaml_env_tag==0.1
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
toml==0.10.2
tomli==2.0.1
tomli_w==1.0.0
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.11.8
uvloop==0.16.0
watchdog==2.1.9
websockets==8.1
zipp==3.8.1
